requirements.txt
pwntools
Hellf
click
